Implement `.getGeneratedKeys[R]` (#9)
I hardcoded it into the base `DbApi`/`Query` APIs, rather than as a
dialect. Sqlite doesn't support it, but it is a builtin part of the JDBC
interface, so it's a bit on the fence which side it should go. But the
fact that it does need special support in the core `DbApi`
implementation makes me lean towards treating it as a builtin

We need to pass an explicit `[R]` type parameter to `.getGeneratedKeys`
because our `Table` model is not rich enough to tell us what the auto
generated primary keys of the table are. This is in contrast to richer
models like SLICK which do contain this information. We can streamline
this later if necessary, but passing in `[Int]` or whatever is not a
huge hardship.