Enhanced Developer Documentation. Docker Setup and Readability Improvements (#10)
👀 
I noticed that the developer documentation section lacked explicit
instructions on setting up **Docker**, a crucial step for running unit
tests and ensuring a seamless development environment. Recognizing the
importance of this step for new and existing developers, I have added
Docker setup instructions to the "Precondition" section of the README.

🆙 💅🏻 
In addition to integrating Docker setup instructions, I took the
opportunity to improve the overall readability of the README file. These
improvements include better formatting, clearer descriptions of commands
and modules, and a more organized presentation of content. This should
make the documentation more **accessible and easier** to follow,
especially for developers who are new to the project.

Co-authored-by: haskiindahouse <>