Schema support (#23)
Quick fix for this issue
https://github.com/com-lihaoyi/scalasql/issues/14

I got a macro error when I made `schemaName: Option[String]`, so to
avoid that I used a `schemaName: String` with a default empty string.

I was only concerned with getting the `FROM` use case to work, but I can
certainly add more tests to try other cases, just let me know. Also let
me know if there's another place that needs a change for this to work
everywhere.

No support for MySQL for now.