Fix option null values failing to put (#42)
This PR fixes the error described in
https://github.com/com-lihaoyi/scalasql/issues/41

Additionally it fixes it for Enum values on Postgres (which I'm fairly
sure had the same error)

It appears keeping record of the JDBCType on TypeMappers is not
necessary anymore, but nonetheless I kept the value there to touch as
little as possible. I did update the DocString.