bump scala3 version and revert scala3 work around (#66)
In #11 a workaround to issue
https://github.com/scala/scala3/issues/19436 was introduced.

Meanwhile the issue was fixed and backported to scala 3.3.5 and 3.5.1.
However we can't use 3.3.5 because we need to "match on higher kinded
types" in the macro and that is only available since scala 3.4.x (SIP-53
I think).

So I decide to bump to scala 3.6.2 and remove the workaround